API

The following functions constitute the public interface of the MultiplexerPaneDisplay package:

MultiplexerPaneDisplay.enable — Function

Enable display via MultiplexerPaneDisplay

MultiplexerPaneDisplay.enable(;
    multiplexer = :tmux,
    target_pane,  # mandatory keyword argument
    verbose = true,
    tmpdir = mktempdir(),
    mux_bin = <"tmux" | "wezterm">
    nrows = 1,
    clear = MultiplexerPaneDisplay.needs_clear(Val(multiplexer)),
    redraw_previous = (clear ? (nrows - 1) : 0),
    imgcat = "",
    dry_run = false,
    only_write_files = false,
    use_filenames_as_title = true,
    sleep_secs = <automatic>,
)

enables Julia's builtin display system to show objects that have an image/png or image/jpeg representation the pane of a terminal multiplexer pane, relying on a sutiable terminal emulator's ability to render images. This works by writing the data to a file in a temporary directory and then calling a suitable imgcat program to show that file in the terminal.

Arguments

multiplexer: The terminal multiplex to use for displaying graphics. One of :tmux, :wezterm.

target_pane: A specification for which multiplexer pane to use as a display. If the multiplexer is :tmux, this can either be an integer indicating a pane in the current tmux window, or a specification of the form "session:window.pane" to use a pane in an arbitrary window of a running tmux session. See The pane number can be obtained by running tmux display -p '#{pane_index}' in a particular pane (or via the C-b q shortcut) . If the multiplexer is :wezterm, this should be the WezTerm pane id, which can be obtained from the WEZTERM_PANE environment variable inside that pane.

mux_bin: The executable for the multiplexer, "tmux" if multiplexer is :tmux, and "wezterm" if multiplexer is :wezterm. This can be set to an absolute path if the executable is not in the shell's PATH.

verbose: If true, show information about the display that is being activated.

tempdir: The directory to which to write the temporary files for each invocation of display. The files will be consecutively numbered, e.g., 001.png, 002.png, etc. The default mktempdir() results in a directory that is automatically deleted on exit. Passing an existing directory will cause the image files to persist in that folder.

imgcat: The command to be used for displaying an image file. It should contain the placeholders {height}, {width}, and {file} The {height} and {width} will be the number of rows/columns in the terminal reserved for drawing the image. These are determined internally based on the size of the pane, the value of nrows, and other factors. The {file} is the absolute path to the temporary image file. If given as an empty string (default), MultiplexerPaneDisplay will attempt to find an available executable and choose parameters based on heuristics. It will attempt to use either the wezterm imgcat command or iTerm's imgcat script.

nrows: The number of image rows that should fit into the pane.

clear: Whether to issue a clear command to the target pane before display. By default, this is chosen based on the multiplexer (true for :tmux, false for :wezterm)

redraw_previous: If set to an integer > 1, for each call to display, first redraw the given number of previous images. This accounts for the lack of scrollback in tmux, allowing to compare the current image with previous images.

dry_run: If true, to not write any files and do not display any images. To be used for debugging in combination with ENV["JULIA_DEBUG"] = MultiplexerPaneDisplay

only_write_files: If true, write image files to tempdir, but to not display them in the target pane. This allows to use MultiplexerPaneDisplay without an actual multiplexer running, under the assumption that the image files can be viewed independently.

use_filenames_as_title: If true, print the filename of the temporary file before rendering the image. Note that more descriptive titles can be used by calling MultiplexerPaneDisplay.display.

sleep_secs: The number of seconds to sleep after drawing each image. This defaults to a heuristic value. If multiplexer=:tmux, and target_pane is a pane in the current window, MultiplexerPaneDisplay needs to actively switch to the target pane, issue the imgcat command, wait for the image to render, and then switch back to the current pane. If sleep_secs to too short so that switching back to the current pane happens before the rendering is complete, the image will be shown in the wrong place. Also, when using redraw_previous to draw multiple images, there may need be be a delay for on image to finish rendering before issuing the command to draw the next image. Typical values are on the order of 0.2-0.5 seconds.

MultiplexerPaneDisplay.enabled — Function

Check whether MultiplexerPaneDisplay is active.

MultiplexerPaneDisplay.enabeld(; verbose=true)

returns true if the currently active display (Base.Multimedia.displays[end]) is provided by MultiplexerPaneDisplay. If verbose=true, also show some information about the configuration of the display.

MultiplexerPaneDisplay.set_options — Function

Modify an active display.

MultiplexerPaneDisplay.set_options(; verbose, kwargs...)

overrides the keyword arguments original given in MultiplexerPaneDisplay.enable for the currently active display for all future invocations of display.

MultiplexerPaneDisplay.display — Function

Show an object with a custom title and other options.

MultiplexerPaneDisplay.display(x; title="", kwargs...)

shows x (which must have a image/png or image/jpeg representation) on the current display, after printing the given title (in lieu of the filename if the display was set up with use_filenames_as_title = true). Additional keyword arguments can be clear, nrows, redraw_previous, imgcat, and use_filenames_as_title, and temporarily override the corresponding options in the display.

Note that this is a different function than Base.display, which does not take keyword arguments.

MultiplexerPaneDisplay.disable — Function

Disable using MultiplexerPaneDisplay as the current display.

MultiplexerPaneDisplay.disable(; verbose = true)

disables the current display if it is managed by MultiplexerPaneDisplay. With verbose=true, also print some information about the settings of the deactivated system, or a warning if the current display is not managed by MultiplexerPaneDisplay.